# openapi.yaml
openapi: 3.0.3
info:
  title: Middle Meetup API
  version: 0.1.0
paths:
  /health:
    get:
      responses:
        "200":
          description: Server health.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/meeting/estimate:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRequest"
      responses:
        "200":
          description: Matrix-based meeting estimate.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingEstimate"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/meeting/routes:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingRequest"
      responses:
        "200":
          description: Detailed meeting routes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingRoutes"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/meeting/area:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MeetingAreaRequest"
      responses:
        "200":
          description: Reachable meeting areas for participants.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeetingArea"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/meeting/destinations/search:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationSearchRequest"
      responses:
        "200":
          description: Ranked meetup destinations matching the query and constraints.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DestinationSearchResponse"
        "400":
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/locations/search:
    get:
      parameters:
        - $ref: "#/components/parameters/LocationQuery"
        - $ref: "#/components/parameters/Lat"
        - $ref: "#/components/parameters/Lng"
        - name: radiusMeters
          in: query
          schema:
            type: integer
            format: int32
            default: 10000
      responses:
        "200":
          description: Matching locations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationList"
        "400":
          description: Invalid search.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/locations/autocomplete:
    get:
      parameters:
        - $ref: "#/components/parameters/LocationQuery"
        - $ref: "#/components/parameters/Lat"
        - $ref: "#/components/parameters/Lng"
      responses:
        "200":
          description: Location suggestions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationList"
        "400":
          description: Invalid autocomplete query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/locations/reverse:
    get:
      parameters:
        - $ref: "#/components/parameters/Lat"
        - $ref: "#/components/parameters/Lng"
      responses:
        "200":
          description: Closest known location.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Location"
        "400":
          description: Invalid reverse geocode query.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    LocationQuery:
      name: query
      in: query
      required: true
      schema:
        type: string
        minLength: 1
    Lat:
      name: lat
      in: query
      required: true
      schema:
        type: number
        format: double
    Lng:
      name: lng
      in: query
      required: true
      schema:
        type: number
        format: double
  schemas:
    Coord:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
    BBox:
      type: object
      required: [minLat, minLng, maxLat, maxLng]
      properties:
        minLat:
          type: number
          format: double
        minLng:
          type: number
          format: double
        maxLat:
          type: number
          format: double
        maxLng:
          type: number
          format: double
    Polygon:
      type: object
      required: [type, coordinates]
      properties:
        type:
          type: string
          enum: [Polygon]
        coordinates:
          type: array
          items:
            type: array
            items:
              type: array
              items:
                type: number
                format: double
    Location:
      type: object
      required: [name, address, coord]
      properties:
        name:
          type: string
        address:
          type: string
        coord:
          $ref: "#/components/schemas/Coord"
    Participant:
      type: object
      required: [name, location]
      properties:
        name:
          type: string
        location:
          $ref: "#/components/schemas/Location"
    MeetingConstraints:
      type: object
      properties:
        maxDurationSeconds:
          type: integer
          format: int32
        radiusMeters:
          type: integer
          format: int32
        sortBy:
          type: string
          enum: [fairest, fastest, lowestMaxTime]
    ParticipantEstimate:
      type: object
      required: [participantName, location, distanceMeters, durationSeconds]
      properties:
        participantName:
          type: string
        location:
          $ref: "#/components/schemas/Location"
        distanceMeters:
          type: integer
          format: int32
        durationSeconds:
          type: integer
          format: int32
    MeetingEstimate:
      type: object
      required: [destination, participants, averageDurationSeconds, totalDurationSeconds, durationSpreadSeconds, maxDurationSeconds]
      properties:
        destination:
          $ref: "#/components/schemas/Location"
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantEstimate"
        averageDurationSeconds:
          type: integer
          format: int32
        totalDurationSeconds:
          type: integer
          format: int32
        durationSpreadSeconds:
          type: integer
          format: int32
        maxDurationSeconds:
          type: integer
          format: int32
    ParticipantArea:
      type: object
      required: [participantName, area]
      properties:
        participantName:
          type: string
        area:
          $ref: "#/components/schemas/Polygon"
    MeetingArea:
      type: object
      required: [participants]
      properties:
        participants:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantArea"
        intersection:
          $ref: "#/components/schemas/Polygon"
        bbox:
          $ref: "#/components/schemas/BBox"
    DestinationCandidate:
      type: object
      required: [location, estimate, score]
      properties:
        location:
          $ref: "#/components/schemas/Location"
        estimate:
          $ref: "#/components/schemas/MeetingEstimate"
        score:
          type: integer
          format: int32
    DestinationSearchResponse:
      type: object
      required: [query, destinations]
      properties:
        query:
          type: string
        area:
          $ref: "#/components/schemas/MeetingArea"
        destinations:
          type: array
          items:
            $ref: "#/components/schemas/DestinationCandidate"
    MeetingRoutes:
      type: object
      required: [destination, routes]
      properties:
        destination:
          $ref: "#/components/schemas/Location"
        routes:
          type: array
          items:
            $ref: "#/components/schemas/ParticipantRoute"
    ParticipantRoute:
      type: object
      required: [participantName, route]
      properties:
        participantName:
          type: string
        route:
          $ref: "#/components/schemas/RouteEstimate"
    RouteEstimate:
      type: object
      required: [from, to, distanceMeters, durationSeconds, geometry]
      properties:
        from:
          $ref: "#/components/schemas/Location"
        to:
          $ref: "#/components/schemas/Location"
        distanceMeters:
          type: integer
          format: int32
        durationSeconds:
          type: integer
          format: int32
        geometry:
          $ref: "#/components/schemas/RouteGeometry"
        steps:
          type: array
          items:
            $ref: "#/components/schemas/RouteStep"
        bbox:
          type: array
          items:
            type: number
            format: double
    RouteGeometry:
      type: object
      properties:
        polyline:
          type: string
        points:
          type: array
          items:
            $ref: "#/components/schemas/Coord"
    RouteStep:
      type: object
      required: [instruction, distanceMeters, durationSeconds]
      properties:
        instruction:
          type: string
        name:
          type: string
        distanceMeters:
          type: integer
          format: int32
        durationSeconds:
          type: integer
          format: int32
        type:
          type: integer
          format: int32
        wayPoints:
          type: array
          items:
            type: integer
            format: int32
        start:
          $ref: "#/components/schemas/Coord"
        end:
          $ref: "#/components/schemas/Coord"
    MeetingRequest:
      type: object
      required: [participants, destination]
      properties:
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        destination:
          $ref: "#/components/schemas/Location"
    MeetingAreaRequest:
      type: object
      required: [participants, constraints]
      properties:
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        constraints:
          $ref: "#/components/schemas/MeetingConstraints"
    DestinationSearchRequest:
      type: object
      required: [participants, query, constraints]
      properties:
        participants:
          type: array
          items:
            $ref: "#/components/schemas/Participant"
        query:
          type: string
        constraints:
          $ref: "#/components/schemas/MeetingConstraints"
        limit:
          type: integer
          format: int32
    LocationList:
      type: array
      items:
        $ref: "#/components/schemas/Location"
    ErrorResponse:
      type: object
      required: [error, code, details]
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: string
    HealthResponse:
      type: object
      required: [ok, message]
      properties:
        ok:
          type: boolean
        message:
          type: string